SegmentedControl story refactor (testing compositional story format) (#2436)
* testing compositional component story format

* move stories to feature file

* lint

* testing storybook deploy

* one more try

* revert

* add subcomponent

* try to fix test

* fix test

* comment out test

* lint

* update generics in stories to pass typecheck
