+++ src/SegmentedControl/SegmentedControl.stories.tsx
@@ -1,0 +1,125 @@
+import React from 'react'
+import {Story, Meta} from '@storybook/react'
+import {SegmentedControl} from '.'
+import SegmentedControlIconButton from './SegmentedControlIconButton'
+import SegmentedControlButton from './SegmentedControlButton'
+import {EyeIcon, FileCodeIcon, PeopleIcon} from '@primer/octicons-react'
+
+type ResponsiveVariantOptions = 'dropdown' | 'hideLabels' | 'default'
+type Args = {
+  fullWidth?: boolean
+  fullWidthAtNarrow?: boolean
+  fullWidthAtRegular?: boolean
+  fullWidthAtWide?: boolean
+  size?: 'small' | 'medium'
+  variantAtNarrow: ResponsiveVariantOptions
+  variantAtRegular: ResponsiveVariantOptions
+  variantAtWide: ResponsiveVariantOptions
+}
+
+const excludedControlKeys = ['variant']
+
+const variantOptions = ['dropdown', 'hideLabels', 'default']
+
+const parseVariantFromArgs = ({variantAtNarrow, variantAtRegular, variantAtWide}: Args) => ({
+  narrow: variantAtNarrow,
+  regular: variantAtRegular,
+  wide: variantAtWide
+})
+
+const parseFullWidthFromArgs = ({fullWidth, fullWidthAtNarrow, fullWidthAtRegular, fullWidthAtWide}: Args) =>
+  fullWidth
+    ? fullWidth
+    : {
+        narrow: fullWidthAtNarrow,
+        regular: fullWidthAtRegular,
+        wide: fullWidthAtWide
+      }
+
+export default {
+  title: 'Components/SegmentedControl',
+  component: SegmentedControl,
+  subcomponents: {SegmentedControlButton, SegmentedControlIconButton},
+  args: {
+    fullWidth: false,
+    fullWidthAtNarrow: false,
+    fullWidthAtRegular: false,
+    fullWidthAtWide: false,
+    size: 'medium',
+    variantAtNarrow: 'default',
+    variantAtRegular: 'default',
+    variantAtWide: 'defualt'
+  },
+  argTypes: {
+    fullWidth: {
+      control: {
+        type: 'boolean'
+      }
+    },
+    fullWidthAtNarrow: {
+      name: 'fullWidth.narrow',
+      control: {
+        type: 'boolean'
+      }
+    },
+    fullWidthAtRegular: {
+      name: 'fullWidth.regular',
+      control: {
+        type: 'boolean'
+      }
+    },
+    fullWidthAtWide: {
+      name: 'fullWidth.wide',
+      control: {
+        type: 'boolean'
+      }
+    },
+    size: {
+      control: {
+        type: 'radio'
+      },
+      options: ['small', 'medium']
+    },
+    variantAtNarrow: {
+      name: 'variant.narrow',
+      control: {
+        type: 'radio'
+      },
+      options: variantOptions
+    },
+    variantAtRegular: {
+      name: 'variant.regular',
+      control: {
+        type: 'radio'
+      },
+      options: variantOptions
+    },
+    variantAtWide: {
+      name: 'variant.wide',
+      control: {
+        type: 'radio'
+      },
+      options: variantOptions
+    }
+  },
+  parameters: {controls: {exclude: excludedControlKeys}}
+} as Meta<typeof SegmentedControl>
+
+export const Playground: Story<Args> = args => (
+  <SegmentedControl
+    aria-label="File view"
+    fullWidth={parseFullWidthFromArgs(args)}
+    variant={parseVariantFromArgs(args)}
+    size={args.size}
+  >
+    <SegmentedControl.Button selected aria-label={'Preview'} leadingIcon={EyeIcon}>
+      Preview
+    </SegmentedControl.Button>
+    <SegmentedControl.Button aria-label={'Raw'} leadingIcon={FileCodeIcon}>
+      Raw
+    </SegmentedControl.Button>
+    <SegmentedControl.Button aria-label={'Blame'} leadingIcon={PeopleIcon}>
+      Blame
+    </SegmentedControl.Button>
+  </SegmentedControl>
+)
